+++ components/org.wso2.bfsi.consent.management.common/src/main/java/org/wso2/bfsi/consent/management/common/config/ConsentManagementConfigurationServiceImpl.java
@@ -36,4 +36,5 @@ public class ConsentManagementConfigurationServiceImpl implements ConsentManagem
     public Map<String, Map<Integer, String>> getConsentAuthorizeSteps() {
         return configParser.getConsentAuthorizeSteps();
     }
+
 }
