+++ components/org.wso2.bfsi.consent.management.common/src/main/java/org/wso2/bfsi/consent/management/common/util/ConsentManagementConstants.java
@@ -46,6 +46,12 @@ public class ConsentManagementConstants {
     public static final String REQUEST_VALIDATOR = "Identity.Extensions.RequestObjectValidator";
     public static final String RESPONSE_HANDLER = "Identity.Extensions.ResponseTypeHandler";
     public static final String CLAIM_PROVIDER = "Identity.Extensions.ClaimProvider";
+    public static final String ENABLE_TRANSPORT_CERT_AS_HEADER = "Identity.ClientTransportCertAsHeaderEnabled";
+    public static final String REMOVE_USER_STORE_DOMAIN_FROM_SUBJECT =
+            "Identity.TokenSubject.RemoveUserStoreDomainFromSubject";
+    public static final String REMOVE_TENANT_DOMAIN_FROM_SUBJECT =
+            "Identity.TokenSubject.RemoveTenantDomainFromSubject";
+    public static final String CONSENT_ID_CLAIM_NAME = "Identity.ConsentIDClaimName";
 
     //Common Constants
     public static final String CONSENT = "consent";
