+++ components/org.wso2.bfsi.identity.extensions/pom.xml
@@ -204,8 +204,8 @@
 
                             javax.validation.*;version="0",
                             org.hibernate.validator.*;version="0",
-                            com.nimbusds.jose;version="${org.wso2.orbit.nimbus.version}",
-                            com.nimbusds.jwt;version="${org.wso2.orbit.nimbus.version}",
+                            com.nimbusds.jose.*;version="${org.wso2.orbit.nimbus.version}",
+                            com.nimbusds.jwt.*;version="${org.wso2.orbit.nimbus.version}",
                             net.minidev.json.*;version="${json-smart.version}",
                             javax.servlet.http; version="${imp.pkg.version.javax.servlet}",
 
@@ -215,6 +215,7 @@
                             org.wso2.carbon.identity.base.*;version="${carbon.identity.framework.version.range}",
                             org.wso2.carbon.identity.application.common.*;version="${carbon.identity.framework.version.range}",
                             org.wso2.carbon.identity.application.mgt;version="${carbon.identity.framework.version.range}",
+                            org.wso2.carbon.identity.application.authentication.framework.*;version="${carbon.identity.framework.version.range}",
                             org.wso2.carbon.identity.core.util;version=""${carbon.identity.framework.version.range}"",
                             org.wso2.carbon.identity.oauth.*;version="${identity.inbound.auth.oauth.version.range}",
                             org.wso2.carbon.identity.oauth2.*;version="${identity.inbound.auth.oauth.version.range}",
@@ -222,8 +223,8 @@
                             org.wso2.carbon.user.api.*;version="${carbon.kernel.version.range}",
                             org.wso2.carbon.user.core.*;version="${carbon.kernel.version.range}",
                             org.wso2.carbon.user.core.service.*;version="${carbon.kernel.version.range}",
+                            org.wso2.carbon.context.*;version="${carbon.kernel.version.range}",
                             org.wso2.carbon.utils.multitenancy;version="${carbon.kernel.version.range}",
-                            *
                         </Import-Package>
                         <Export-Package>
                             !org.wso2.bfsi.identity.extensions.internal,
