--- components/org.wso2.bfsi.identity.extensions/src/main/java/org/wso2/bfsi/identity/extensions/util/ServiceProviderUtils.java
@@ -1,58 +1,0 @@
-/**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.bfsi.identity.extensions.util;
-
-import org.wso2.bfsi.consent.management.common.exceptions.ConsentManagementException;
-import org.wso2.bfsi.consent.management.common.util.Generated;
-import org.wso2.carbon.identity.base.IdentityRuntimeException;
-import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
-import org.wso2.carbon.identity.oauth.common.exception.InvalidOAuthClientException;
-import org.wso2.carbon.identity.oauth2.IdentityOAuth2Exception;
-import org.wso2.carbon.identity.oauth2.util.OAuth2Util;
-
-/**
- * Utility Class for Service Provider related functions.
- */
-public class ServiceProviderUtils {
-
-    /**
-     * Get Tenant Domain String for the client id.
-     * @param clientId the client id of the application
-     * @return tenant domain of the client
-     * @throws ConsentManagementException  if an error occurs while retrieving the tenant domain
-     */
-    @Generated(message = "Ignoring because OAuth2Util cannot be mocked with no constructors")
-    public static String getSpTenantDomain(String clientId) throws ConsentManagementException {
-
-        try {
-            return OAuth2Util.getTenantDomainOfOauthApp(clientId);
-        } catch (InvalidOAuthClientException | IdentityOAuth2Exception e) {
-            throw new ConsentManagementException("Error retrieving service provider tenant domain for client_id: "
-                    + clientId, e);
-        }
-    }
-
-    public static int getTenantIdOfUser(String userId) throws ConsentManagementException {
-        try {
-            return IdentityTenantUtil.getTenantIdOfUser(userId);
-        } catch (IdentityRuntimeException e) {
-            throw new ConsentManagementException("Error retrieving tenant id for user: " + userId, e);
-        }
-    }
-}
